Total reconstruction:
1) reconstructed CMakeLists.exe to support WIN OS simulation and ARM
Build (BUILD_FOR_SIM_ON_WIN_OS flag). The win os simulation port is
located at [flight-computer/sim-port/os-kernel]. Both IMU and Pressure
sensors have their corresponding ports as well as the UART, SPI protocols
and HAL. Lots of functionalities of these three compponents are just stubbed.

2) Datafeeder pipeline for the WIN OS simulation: IMU and Pressure
sensor uses CSV file that is emmited in the system from the datafeeder
thread. CSV file support is enabled through -DSRAD_CSV_FILE_PATH flag in
CMakeLists that is automatically assigned with pathes of the csv files in
[flight-computer/sim-port/sensor-simulation].

3) fixed sensors APIs include pathes

4) incorporating the new configuration system for sensors and other
threads. The idea is that the certain configuratinos are grouped
together into the corresponding structs. While all the structs are
inorporated into the global system configuration struct. This idea also
includes hiding and creating default configuration profiles for sensors
and hiding its internals with regard to configurations. This idea has to
do with the fact that not all configurations needs to be accesssed to
the components. We give access only to those components of the
configuratinos that are required.

5) Removing ref_press and ref_alt cmponent from the sensor
implementation following a single responsibility paradigm. Momed this
component to the event detection task because it is important to know
the reference pressure only when calculating the events.

6) changing timestampts from floats to uint32s

7) Setting up the basic CLI using FreeRTOS-CLI library and incorporating
some of the interfaces from the old version (configure, mem, etc.).
Adding up a support of a syntax check for each option=argument.

8) Adding support of a DEBUG-level print; enabling or disabling with the
flag -DPRINT_DEBUG_LOG.

9) added __cplusplus__ preprocessor directives to support c++ execution

10) added common_ prefix to the common functions for [utilities/common.h]

+++ flight-computer/memory-management/memory_manager.h
@@ -35,7 +35,7 @@ typedef struct
 typedef union
 {
     struct {
-        float timestamp;
+        uint32_t timestamp;
         float pressure;
         float temperature;
     };
@@ -60,7 +60,7 @@ typedef union
 {
     uint8_t updated;
     struct {
-        float timestamp;
+        uint32_t timestamp;
         ContinuityStatus status;
     };
     uint8_t bytes[sizeof(float) * 1 +
@@ -77,7 +77,7 @@ typedef struct
 typedef union
 {
     struct {
-        float timestamp;
+        uint32_t timestamp;
         FlightEventStatus status;
     };
     uint8_t bytes[sizeof(float) * 1 +
@@ -94,7 +94,7 @@ typedef struct
 
 typedef struct
 {
-    float timestamp;
+    uint32_t timestamp;
     IMUData inertial;
     PressureData pressure;
     Continuity continuity;
